Custom events were not fired. Event that are now managed are:
VPReady: When the video is ready to start playing
VPStart: When the video start playing (time is passed on the event object)
VPProgress: fired every 200 ms when the video is in progress.
VPPlay: Each time the video change from paused to play.
VPPause: When the video pause (time is passed on the event object)
VPEnd: When video end; you can retrieve the time on the event object.
VPFullScreenStart: When video goes full screen modality
VPFullScreenEnd: When video exit full screen modality
VPMuted: When video is muted
VPUnmuted: When video is unmuted
VPTime: Fired every seconds (time is passed on the event object: e.time)